backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
language-tool-python==2.7.1
